changes done in the add Lead,Login by User,Admin and Role based API.

+++ src/main/java/ai/rnt/crm/dto/LeadDto.java
@@ -27,7 +27,7 @@ public class LeadDto {
 
 	private String businessCard;
 
-	private Float budgetAmount;
+	private Double budgetAmount;
 
 	private Integer assignTo;
